Renamed RouterItem classes.
